refactor: Racing Game を features/ 配下に分割
- RacingGamePage.tsx (1,820行) を src/features/racing-game/ 配下に分割
- types.ts, constants.ts, utils.ts, audio.ts, entities.ts, track.ts,
  renderer.ts, game-logic.ts, hooks.ts, components/VolumeControl.tsx,
  RacingGame.tsx, index.ts の12ファイルに分離
- RacingGamePage.tsx を薄いラッパー (6行) に変更
- utils, entities, track, game-logic のユニットテスト47件を追加
- 既存テスト5件 + 新規47件 = 全52件パス確認済み
