+++ src/features/labyrinth-of-shadows/index.ts
@@ -1,0 +1,19 @@
+export { default as LabyrinthOfShadowsGame } from './LabyrinthOfShadowsGame';
+export type {
+  Difficulty,
+  EntityType,
+  SoundName,
+  Entity,
+  Player,
+  Enemy,
+  Item,
+  Sprite,
+  GameState,
+  HUDData,
+} from './types';
+export { CONFIG, CONTENT } from './constants';
+export { MazeService } from './maze-service';
+export { AudioService } from './audio';
+export { EntityFactory, GameStateFactory } from './entity-factory';
+export { GameLogic } from './game-logic';
+export { Renderer } from './renderer';
