+++ src/features/racing-game/index.ts
@@ -1,0 +1,13 @@
+// Racing Game re-export
+
+export { default as RacingGame } from './RacingGame';
+export type { Point, Checkpoint, StartLine, Particle, Spark, Confetti, Course, Player, Decoration, GameResults } from './types';
+export { Config, Colors, Options, Courses } from './constants';
+export { Utils } from './utils';
+export { SoundEngine } from './audio';
+export { Entity } from './entities';
+export { Track } from './track';
+export { Render, renderDecos } from './renderer';
+export { Logic } from './game-logic';
+export { useInput, useIdle } from './hooks';
+export { VolumeCtrl } from './components/VolumeControl';
